Move lz4 compressor to separate module (#1567)
Some users have problems using gocql/gocql with dep as
dep can't work with v4 dependency.
Moving the lz4 compressor to a separate module.
This is a breaking change, but there probably aren't many
people using the lz4 compressor yet.
