+++ lz4/go.mod
@@ -1,0 +1,8 @@
+module github.com/gocql/gocql/lz4
+
+go 1.16
+
+require (
+	github.com/pierrec/lz4/v4 v4.1.8
+	github.com/stretchr/testify v1.7.0
+)
